+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to the "ollama-autopilot" extension will be documented in this file.
 
-## [Unreleased]() - 2026-02-21
+## [1.1.0](https://github.com/dadul96/ollama-autopilot/releases/tag/v1.1.0) - 2026-02-22
 ### Added
 - Keybinding (ctrl+alt+space) for manual trigger
 - Option to choose between automatic or manual suggestion trigger